+++ plugins/src/airview/find_parameters.py
@@ -1,0 +1,55 @@
+import multi_scale
+import math
+
+def findSumabsSumsqN_row(input, scale1, scale2):
+    '''
+    Calculates the sum of absolute values, sum of squares, and the size of one region.
+    
+    return:
+    list[float]: sum of absolute values, sum of squares, and size.
+    '''
+    sumabs_sumsq_n = [None, None, None]
+    # get regions with constant power - multiscale transform
+    # regions is list of lists
+    # TODO
+    regions = multi_scale.multiscale_detection_getDefaultRegions(input, scale1, scale2)
+
+    for i in range(1, len(regions)):
+        sumabs_sumsq_n[0] += abs(regions.get[i - 1][4] - regions.get[i][4])
+        sumabs_sumsq_n[1] += (regions.get[i - 1][4] - regions.get[i][4]) ** 2
+        sumabs_sumsq_n[2] += 1
+    
+    return sumabs_sumsq_n
+
+def findSumabsSumsqN(input, scale1, scale2):
+    '''
+    Calculates the sum of absolute values, sum of squares, and size
+    
+    return:
+    list[float]: sum of absolute values, sum of squares, and size
+    '''
+    # sumabs_sumsq_n[0] = sum of absolute values
+    # sumabs_sumsq_n[1] = sum of squares
+    # sumabs_sumsq_n = size
+    sumabs_sumsq_n = [None, None, None]
+    for i, row in enumerate(input): # iterate through rows
+        loc_sumabs_sumsq_n = findSumabsSumsqN_row(row, scale1, scale2) # compute the sum of absolute values, sum of squares, and size for each row
+        for j, s in enumerate(sumabs_sumsq_n):
+            s += loc_sumabs_sumsq_n[j] # accumulate results
+    return sumabs_sumsq_n
+
+def findAvgAdjDiffCoarse(input, scale1, scale2):
+    '''
+    Finds the average/standard deviation of adjacent differences in coarse
+	signals NOTE: This method just uses the "default" regions (those defined
+	by the resolution) to construct the coarse signal
+	
+	return:
+    list[float]: mean and standard deviation of data
+    '''
+    mean_stdev = [None, None]
+    sumabs_sumsq_n = findSumabsSumsqN(input, scale1, scale2)
+    mean_stdev[0] = sumabs_sumsq_n[0]/sumabs_sumsq_n[2]
+    mean_stdev[1] = math.sqrt(sumabs_sumsq_n[1]/sumabs_sumsq_n[2] - (mean_stdev[0]**2))
+    return mean_stdev
+
